refactor: move flags() to fast-usdc/tools/
